@maggiebowman6 @jegan002  got the plotting working
Was one straggler file that didn't get a 6 on the end of the name, and messed things up.

I added a bulk plot function to the package, you can reinstall to use (0.1.7).

I also added the script I used to correct all this stuff, it is in "/example_data/batch_process_script.R".