Got a 300,300 fix
@maggiebowman6 Ran it on MaggieExtracts to check it.

The function is fluoro:::mask.300(eem.dir), and will mask 300,300 peaks in all the corrected files it can see in eem.dir. I added it to the batch process script for an example of how to use it.

Because it messes with corrected EEMs, I made it a 'hidden' function. It has to be called with 'fluoro:::' at the beginning (3x":")

Hope that works!